requirements.txt
pillow==1.7.8
requests==1.2.3
